+++ BruteForcePassword/Sources/Controller/Extension.swift
@@ -1,0 +1,23 @@
+//
+//  Extension.swift
+//  BruteForcePassword
+//
+//  Created by Игорь Николаев on 15.02.2023.
+//
+
+import Foundation
+
+extension String {
+    var digits: String { return "0123456789" }
+    var lowercase: String { return "abcdefghijklmnopqrstuvwxyz" }
+    var uppercase: String { return "ABCDEFGHIJKLMNOPQRSTUVWXYZ" }
+    var punctuation: String { return "!\"#$%&'()*+,-./:;<=>?@[\\]^_`{|}~" }
+    var letters: String { return lowercase + uppercase }
+    var printable: String { return digits + letters + punctuation }
+
+    mutating func replace(at index: Int, with character: Character) {
+        var stringArray = Array(self)
+        stringArray[index] = character
+        self = String(stringArray)
+    }
+}
